revert .gitignore
Since the .ipynb files are already under version control, revising the .gitignore doesn't "just work" and it's possible we want to add more .ipynb files to the repo at some point. So for now, if you change it and you don't want to keep those changes, use the 'checkout' workflow in the Wiki.